@@ -3,7 +3,7 @@
 
     using FluentValidation;
 
-    using ProjectManager.Common.Models;
+    using ProjectManager.Data.Models;
 
     /// <summary>
     ///     The validator for the project.
