@@ -2,7 +2,6 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.Diagnostics.CodeAnalysis;
     using System.Threading;
     using System.Threading.Tasks;
 
@@ -12,12 +11,11 @@
     using Moq;
 
     using ProjectManager.Business.Services;
-    using ProjectManager.Common.Models;
+    using ProjectManager.Data.Models;
     using ProjectManager.MVC.Controllers;
 
     using Xunit;
 
-    [ExcludeFromCodeCoverage]
     public sealed class ProjectsControllerTests
     {
         private readonly MockRepository mockRepository;
